Conflicts:
	src/main/java/com/clevel/selos/businesscontrol/PrescreenBusinessControl.java
	src/main/java/com/clevel/selos/controller/PrescreenChecker.java
	src/main/java/com/clevel/selos/controller/PrescreenMaker.java
	src/main/java/com/clevel/selos/dao/master/UserDAO.java
	src/main/java/com/clevel/selos/model/view/PrescreenCollateralView.java
	src/main/java/com/clevel/selos/transform/BankTransform.java
	src/main/java/com/clevel/selos/transform/BizProductDetailTransform.java
	src/main/java/com/clevel/selos/transform/LoanTypeTransform.java
	src/main/java/com/clevel/selos/transform/PrescreenCollateralTransform.java
	src/main/java/com/clevel/selos/util/DateTimeUtil.java
	src/main/webapp/resources/js/custom.js
	src/main/webapp/site/dialog/prescreenCustomerDialog.xhtml
	src/main/webapp/site/dialog/prescreenProposeCollateralDialog.xhtml
	src/main/webapp/site/prescreenMaker.xhtml

+++ src/main/java/com/clevel/selos/dao/working/DBRDetailDAO.java
@@ -6,12 +6,12 @@ import org.slf4j.Logger;
 
 import javax.inject.Inject;
 
-public class DBRDetailDAO extends GenericDAO<DBRDetail, Integer> {
+public class DBRDetailDAO extends GenericDAO<DBRDetail, Long> {
     @Inject
     private Logger log;
 
     @Inject
-    public DBRDetailDAO(){
+    public DBRDetailDAO() {
 
     }
 }
